modulo usuario atualizado e rabase aplicado
