+++ src/main/kotlin/br/ufrn/imd/obama/usuario/infrastructure/adapter/UsuarioGatewayAdapter.kt
@@ -1,0 +1,19 @@
+package br.ufrn.imd.obama.usuario.infrastructure.adapter
+
+import br.ufrn.imd.obama.usuario.domain.gateway.UsuarioGateway
+import br.ufrn.imd.obama.usuario.infrastructure.mapper.toEnitty
+import br.ufrn.imd.obama.usuario.domain.model.Usuario
+import br.ufrn.imd.obama.usuario.infrastructure.entity.UsuarioEntity
+import br.ufrn.imd.obama.usuario.infrastructure.mapper.toModel
+import br.ufrn.imd.obama.usuario.infrastructure.repository.UsuarioRepository
+import org.springframework.stereotype.Service
+
+@Service
+class UsuarioGatewayAdapter (
+    private val usuarioRepository: UsuarioRepository
+): UsuarioGateway {
+    override fun salvarUsuario(usuario: Usuario): Usuario {
+        val usuarioEntity: UsuarioEntity = usuario.toEnitty()
+        return usuarioRepository.save(usuarioEntity).toModel()
+    }
+}
