+++ src/main/kotlin/br/ufrn/imd/obama/usuario/infrastructure/repository/UsuarioRepository.kt
@@ -1,6 +1,6 @@
-package br.ufrn.imd.obama.oa.infrastructure.repository
+package br.ufrn.imd.obama.usuario.infrastructure.repository
 
-import br.ufrn.imd.obama.oa.infrastructure.entity.UsuarioEntity
+import br.ufrn.imd.obama.usuario.infrastructure.entity.UsuarioEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
