+++ src/main/kotlin/br/ufrn/imd/obama/usuario/infrastructure/resource/exchange/UsuarioResponse.kt
@@ -1,7 +1,7 @@
-package br.ufrn.imd.obama.oa.infrastructure.resource.exchange
+package br.ufrn.imd.obama.usuario.infrastructure.resource.exchange
 
-import br.ufrn.imd.obama.oa.domain.enums.Role
-import br.ufrn.imd.obama.oa.domain.enums.TipoCadastro
+import br.ufrn.imd.obama.usuario.domain.enums.Role
+import br.ufrn.imd.obama.usuario.domain.enums.TipoCadastro
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
